Add horus requirements magic command
